Refactored a metric fuck-ton (READ DESCRIPTION)
The following scripts are now obsolete (not yet deleted)
-PlayerMovement.cs
-PlayerControls.cs
-Incontrol

They have been replaced with:
-BasePlayer.cs (abstract class)
-Player.cs
-Player1.cs (inherited class)
-Player2.cs (inherited class)

The following scripts have been modified to reflect these changes:
-BallMovement.cs
-EndOfRound.cs
-LoadSelection.cs

NOTE: Racket Toss mechanic does NOT currently work with the refactored
system. It will be implemented into the abstract and inherited classes
in the next push.